+++ tests/testthat/test-odbc-connection.R
@@ -7,14 +7,6 @@ test_that("handles simple inputs", {
   expect_equal(build_connection_string(foo = "1", bar = "2"), "foo=1;bar=2")
 })
 
-test_that("escapes if needed", {
-  expect_equal(build_connection_string(foo = "*"), "foo={*}")
-  # Already wrapped
-  expect_equal(build_connection_string(foo = "{*}"), "foo={*}")
-  # Respects AsIs
-  expect_equal(build_connection_string(foo = I("*")), "foo=*")
-})
-
 test_that("combines with existing .connection string", {
   expect_equal(build_connection_string("x=1"), "x=1")
   expect_equal(build_connection_string("x=1", foo = "1"), "x=1;foo=1")
@@ -34,4 +26,36 @@ test_that("errors about case-insensitve arguments", {
     build_connection_string(xa = 1, xA = 1, XA = 1)
     build_connection_string(xa = 1, xA = 1, xb = 1, xB = 1)
   })
+})
+
+test_that("warns if your values might need quoting", {
+  expect_snapshot(build_connection_string(foo = "f{o"))
+})
+
+test_that("correctly detects values that need escaping", {
+  # I() opts-out
+  expect_false(needs_quoting(I("f{o")))
+
+  # already quoted
+  expect_false(needs_quoting("'f{o'"))
+  expect_false(needs_quoting('"f{o"'))
+  expect_false(needs_quoting("{f{o}"))
+
+  # no special values
+  expect_false(needs_quoting("foo"))
+
+  # has a special character
+  expect_true(needs_quoting("foo "))
+  expect_true(needs_quoting("foo{"))
+  expect_true(needs_quoting("foo}"))
+  expect_true(needs_quoting("foo="))
+  expect_true(needs_quoting("foo;"))
+})
+
+test_that("automatically picks correct quote type", {
+  expect_equal(quote_value("x"), I('"x"'))
+  expect_equal(quote_value("x'"), I('"x\'"'))
+  expect_equal(quote_value("x'"), I('"x\'"'))
+
+  expect_snapshot(quote_value("'\""), error = TRUE)
 })